1.0.0.7 +fixed minor bugs +add button visibly/hidden column ID (by default: hidden).

+++ ListOfBasesUsers/DirFile.cs
@@ -27,23 +27,21 @@ namespace ListOfBasesUsers
 
         internal List<string> GetListUsers()
         {
-
             List<string> list = new List<string>();
 
             DirectoryInfo dirInfo = new DirectoryInfo(DefaultValues.pathUsers);
 
             foreach (DirectoryInfo item in dirInfo.GetDirectories())
             {
-
-                if (
+                if (!(
                     (item.Attributes & FileAttributes.Hidden) == FileAttributes.Hidden
-                    || 
+                    ||
                     (item.Attributes & FileAttributes.System) == FileAttributes.System
-                    )
-                    continue;
+                    ))
 
-                list.Add(item.Name);
-
+                {
+                    list.Add(item.Name);
+                }
             }
 
             list.Sort(delegate(string x, string y)
@@ -53,7 +51,6 @@ namespace ListOfBasesUsers
             );
 
             return list;
-
         }
 
         internal ulong GetDirSize() => CalculateSize(Path);
@@ -62,9 +59,11 @@ namespace ListOfBasesUsers
 
         internal string GetSizeFormat(ulong DirSize)
         {
+            if (DirSize == 0)
+                return "";
 
-            if (DirSize < 1024)
-                return (DirSize).ToString("F0") + " bytes";
+            else if (DirSize < 1024)
+                return (DirSize).ToString("F0") + " B";
 
             else if ((DirSize >> 10) < 1024)
                 return (DirSize / (float)1024).ToString("F1") + " KB";
@@ -83,7 +82,6 @@ namespace ListOfBasesUsers
 
             else
                 return ((DirSize >> 50) / (float)1024).ToString("F0") + " EB";
-
         }
 
         internal void OpenDirectory()
@@ -93,7 +91,6 @@ namespace ListOfBasesUsers
 
         internal Process OpenDirectory(bool returnProcess)
         {
-
             Process process = new Process()
             {
                 EnableRaisingEvents = true,
@@ -101,7 +98,6 @@ namespace ListOfBasesUsers
             };
 
             return process;
-
         }
 
         internal void DeleteCatalogCache()
@@ -154,42 +150,44 @@ namespace ListOfBasesUsers
 
         private void DeleteSubDir(string path)
         {
-
             DirectoryInfo dirInfo = new DirectoryInfo(path);
-            foreach (FileInfo file in dirInfo.GetFiles())
+            if (dirInfo.Exists)
             {
-                try
+                foreach (FileInfo file in dirInfo.GetFiles())
                 {
-                    file.Delete();
+                    try
+                    {
+                        file.Delete();
+                    }
+                    catch (Exception)
+                    {
+                    }
                 }
-                catch (Exception)
+                foreach (DirectoryInfo currentSubDir in dirInfo.GetDirectories())
                 {
+                    try
+                    {
+                        DeleteSubDir(currentSubDir.FullName);
+                        currentSubDir.Delete(true);
+                    }
+                    catch (Exception)
+                    {
+                    }
                 }
             }
-            foreach (DirectoryInfo currentSubDir in dirInfo.GetDirectories())
-            {
-                try
-                {
-                    DeleteSubDir(currentSubDir.FullName);
-                    currentSubDir.Delete(true);
-                }
-                catch (Exception)
-                {
-                }
-            }
-
         }
 
         private ulong CalculateSize(string path)
         {
             ulong size = 0;
+            if (new DirectoryInfo(path).Exists)
+            {
+                foreach (string files in Directory.GetFiles(path))
+                    size += (ulong)new FileInfo(files).Length;
 
-            foreach (string files in Directory.GetFiles(path))
-                size += (ulong)new FileInfo(files).Length;
-
-            foreach (string dir in Directory.GetDirectories(path))
-                size += CalculateSize(dir);
-
+                foreach (string dir in Directory.GetDirectories(path))
+                    size += CalculateSize(dir);
+            }
             return size;
         }
